requirements.txt
streamlit>=1.37
numpy>=1.24
pypdf>=4.2
google-generativeai>=0.8
